fixed registration bug

+++ src/main/java/net/proselyte/springsecurityapp/model/Role.java
@@ -5,6 +5,7 @@ import java.util.Set;
 
 /**
  * Created by tramon on 26.01.2017.
+ *
  */
 
 
@@ -14,7 +15,7 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
 
     @Column(name = "name")
     private String name;
@@ -25,11 +26,11 @@ public class Role {
     public Role() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
