+++ src/main/java/net/proselyte/springsecurityapp/model/User.java
@@ -4,7 +4,8 @@ import javax.persistence.*;
 import java.util.Set;
 
 /**
- * Created by tramo_000 on 26.01.2017.
+ * Created by tramon on 26.01.2017.
+ *
  */
 
 @Entity
@@ -13,7 +14,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
 
     @Column(name = "username")
     private String username;
@@ -33,11 +34,11 @@ public class User {
     public User() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
